Make A_Sprite its own Struct
The idea is that instead of being just an alias for SDL_Texture*, we will now make it a struct to store
more data about the Sprite. In future commits, this will be made into a SpriteSheet, so we will need to store things
like width, height and number of frames.