Add Velocity component
Add a new component type to represent velocity. It gets added to the
position of the node every frame. For this, also introduced the
concept of delta time, which is the number of seconds, represented
as float, since the last frame.

Flappy now falls down slowly. That is because the velocity is static
and there is no accelaration.